2lab/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pydantic-settings
email-validator
python-multipart
